fix: Items in list were left expanded when offscreen
This commit changes the whole animation scheme. To properly work with the list, we have to send update notifications to it. So we moved the animation logic to inside a DiffUtils implementation so it can properly dictates which items will be updated. With this, even if the ViewHolder is not visible, the update will be properly handled in the list.